feat : 구간 추가 기능 구현 todo: 거리검증, 노선 역 목록 반환

+++ src/main/java/wooteco/subway/line/dto/LineRequest.java
@@ -6,12 +6,12 @@ public class LineRequest {
     private Long upStationId;
     private Long downStationId;
     private int distance;
-    private Long extraFare;
+    private int extraFare;
 
     public LineRequest() {
     }
 
-    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance, Long extraFare) {
+    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance, int extraFare) {
         this.name = name;
         this.color = color;
         this.upStationId = upStationId;
@@ -40,7 +40,7 @@ public class LineRequest {
         return distance;
     }
 
-    public Long getExtraFare() {
+    public int getExtraFare() {
         return extraFare;
     }
